requirements.txt
Flask==3.1.0
flask_socketio==5.5.1
python-dotenv==1.0.1
